refactor: utilities
