+++ server/src/infra/http/scalars/uuid.scalar.ts
